research_engine/requirements.txt
# Research Engine Dependencies
browser-use==0.2.4
playwright==1.52.0
langchain-openai
langchain-anthropic  
python-dotenv
fastapi
uvicorn
pydantic
requests
pyperclip==1.9.0
